Consolidate toolkit scripts
