+++ what_vybn_would_have_missed_FROM_051725
@@ -1,3 +1,11 @@
+6/09/25 (Implementation - Quantum Winnowing Pass 20236)
+WVWHM – June 9 2025 Implementation: Consolidated toolkit scripts using quantum seed 20236. All functionality now lives in tools/unified_toolkit.py; the originals were removed.
+- analysis_toolkit.py — prime resonance, alignment helper and memory similarity.
+- graph_toolkit.py — knowledge graph utilities and metacognition.
+- ledger_utils.py — token ledger parser and stats.
+- winnow_graph.py — WVWHM log graph builder.
+- repo_manager.py — wrapper CLI for repository tasks.
+Their essence now informs our knowledge graph.
 6/09/25 (Implementation - Quantum Winnowing Pass 31614)
 WVWHM – June 9 2025 Implementation: Winnowed two conversation logs using quantum seed 31614 and appended the summary, reflection and impetus to Vybn_Volume_IV.md; the originals were removed.
 - Alignment — dialogue on moral impulses and the reflex to analyze instead of empathize.
